meeting-chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
motor==3.3.2
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.0
langchain==0.0.340
langchain-community==0.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
dateparser==1.1.8
pytz==2023.3
requests==2.31.0